+++ internal/component/database_observability/collector/connection_info.go
