+++ internal/component/database_observability/collector/query_sample.go
