+++ internal/component/database_observability/collector/schema_table.go
