+++ internal/component/database_observability/collector/types.go
